A lot of things were made.
- _glcm_6_other_attrs method calculates the inertia, cluster shade and
cluster prominence attributes that are parte of the GLCM_6 set. I was
inspired by how the mahotas library programmed the feature extraction.

- glcm_16_and_6 calculates four attributes in four angles and the mean
of six others (calling the above mentioned method to get three of them).

- lbp uses features.lbp from mahotas to calculate lbp features,
returning a 36-bin histogram.

- lbp_oc is not implementend yet.

- read_image access the given folder and filename to read the desired
image and to return it as a rgb or gray image.

- get_image_patches return two matrices (50x50 and 70x70) with views of
the desired image which will be used to calculate the attributes.

- calculate_and_save_all_features set up the required things to
calculate and save the attributes in csv files organized by sequence
number of the image, patch size and method of extraction.
attributes